// src/plot/client/src/gui/JAIAProtobuf.ts
export interface GeographicCoordinate {
    lat?: number
    lon?: number
}
// ...
export interface MissionTask {
    type?: TaskType
    dive?: DiveParameters
    surface_drift?: DriftParameters
}

export enum MissionStart {
    START_IMMEDIATELY = "START_IMMEDIATELY",
    START_ON_COMMAND = "START_ON_COMMAND",
}

export enum MovementType {
    TRANSIT = "TRANSIT",
    REMOTE_CONTROL = "REMOTE_CONTROL",
}

export interface Goal {
    name?: string
    location?: GeographicCoordinate
    task?: MissionTask
}

export interface Recovery {
    recover_at_final_goal?: boolean
    location?: GeographicCoordinate
}

export interface Speeds {
    transit?: number
    stationkeep_outer?: number
}

export interface MissionPlan {
    start?: MissionStart
    movement?: MovementType
    goal?: Goal[]
    recovery?: Recovery
    speeds?: Speeds
}